Properly cast 'time_t' to 'uint64_t' (close #4)
The exact type of 'time_t' is platform dependent and therefore
can't be assumed to be 'uint64_t' all the time.

For example, on 32bit platforms, where 'time_t' is defined as long
(32bit), the compiler will generate the following warning, due
to the arithmetic used in the macro:

ulid_struct.hpp:248:34: warning: shift count >= width of type
[-Wshift-count-overflow]

Force time_t to be parsed as uint64_t.